# Conflicts:
#	api/v1beta2/moduletemplate_types.go

+++ internal/manifest/custom_resource.go
@@ -1,4 +1,4 @@
-package v1beta1
+package manifest
 
 import (
 	"context"
@@ -7,8 +7,8 @@ import (
 	"strings"
 
 	"github.com/kyma-project/lifecycle-manager/api/v1beta2"
+	"github.com/kyma-project/lifecycle-manager/pkg/util"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/api/meta"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime/schema"
@@ -30,15 +30,17 @@ var (
 func PostRunCreateCR(
 	ctx context.Context, skr declarative.Client, kcp client.Client, obj declarative.Object,
 ) error {
-	manifest := obj.(*v1beta2.Manifest)
+	manifest, ok := obj.(*v1beta2.Manifest)
+	if !ok {
+		return nil
+	}
 	if manifest.Spec.Resource == nil {
 		return nil
 	}
-	resource := manifest.Spec.Resource.DeepCopy()
 
-	if err := skr.Create(
-		ctx, resource, client.FieldOwner(declarative.CustomResourceManager),
-	); err != nil && !k8serrors.IsAlreadyExists(err) {
+	resource := manifest.Spec.Resource.DeepCopy()
+	err := skr.Create(ctx, resource, client.FieldOwner(declarative.CustomResourceManager))
+	if err != nil && !k8serrors.IsAlreadyExists(err) {
 		return err
 	}
 
@@ -67,20 +69,22 @@ func PostRunCreateCR(
 func PreDeleteDeleteCR(
 	ctx context.Context, skr declarative.Client, kcp client.Client, obj declarative.Object,
 ) error {
-	manifest := obj.(*v1beta2.Manifest)
+	manifest, ok := obj.(*v1beta2.Manifest)
+	if !ok {
+		return nil
+	}
 	if manifest.Spec.Resource == nil {
 		return nil
 	}
-	resource := manifest.Spec.Resource.DeepCopy()
 
+	resource := manifest.Spec.Resource.DeepCopy()
 	propagation := v1.DeletePropagationBackground
 	err := skr.Delete(ctx, resource, &client.DeleteOptions{PropagationPolicy: &propagation})
-
 	if err == nil {
 		return ErrWaitingForAsyncCustomResourceDeletion
 	}
 
-	if !k8serrors.IsNotFound(err) && !meta.IsNoMatchError(err) {
+	if !util.IsNotFound(err) {
 		return err
 	}
 
@@ -97,13 +101,13 @@ func PreDeleteDeleteCR(
 		return ErrWaitingForAsyncCustomResourceDefinitionDeletion
 	}
 
-	if !k8serrors.IsNotFound(err) && !meta.IsNoMatchError(err) {
+	if !util.IsNotFound(err) {
 		return err
 	}
 
 	onCluster := manifest.DeepCopy()
 	err = kcp.Get(ctx, client.ObjectKeyFromObject(obj), onCluster)
-	if k8serrors.IsNotFound(err) {
+	if util.IsNotFound(err) {
 		return nil
 	}
 	if err != nil {
@@ -120,14 +124,17 @@ func PreDeleteDeleteCR(
 }
 
 func GetModuleCRDName(obj declarative.Object) string {
-	manifest := obj.(*v1beta2.Manifest)
-	if manifest.Spec.Resource != nil {
-		group := manifest.Spec.Resource.GroupVersionKind().Group
-		name := manifest.Spec.Resource.GroupVersionKind().Kind
-		return fmt.Sprintf("%s.%s", getPlural(name), group)
+	manifest, ok := obj.(*v1beta2.Manifest)
+	if !ok {
+		return ""
+	}
+	if manifest.Spec.Resource == nil {
+		return ""
 	}
 
-	return ""
+	group := manifest.Spec.Resource.GroupVersionKind().Group
+	name := manifest.Spec.Resource.GroupVersionKind().Kind
+	return fmt.Sprintf("%s.%s", getPlural(name), group)
 }
 
 func getPlural(moduleName string) string {
